Fix to OpenFlow 1.3 PacketOut in_port() return size

+++ fluid/of13msg.hh
@@ -275,7 +275,7 @@ public:
     bool operator!=(const PacketOut &other) const;
     uint8_t* pack();
     of_error unpack(uint8_t *buffer);
-    uint16_t in_port() {
+    uint32_t in_port() {
         return this->in_port_;
     }
     void in_port(uint32_t in_port) {
